requirements.txt
requests
streamlit
python-dotenv
langchain-core
langchain-openai
